+++ Team6220_PowerPlay/src/main/java/org/firstinspires/ftc/team6220_PowerPlay/testclasses/ConeAndJunctionDetectionPipeline.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.team6220_PowerPlay;
+package org.firstinspires.ftc.team6220_PowerPlay.testclasses;
 
 import org.opencv.core.Core;
 import org.opencv.core.Mat;
